corrected vcl-reload.sh
The previous version seemed to create a weird situation:
the _.vsm of the running instance was removed and overriden by another
one. This prevented varnish to send logs to the shared memory (VSM),
thus prevented varnishncsa and varnishlog to get any output.

This problem was raised each time we called the vcl-reload script in
order to update the varnish configuration…

The TMPDIR was removed due to a problem on newer varnish version who
forced varnish user by default for varnishd runs, preventing it to
access the TMPDIR. Adding the -u root in the command line in vcl-reload
seems to prevent such a behavior and seems to ensure varnish is properly
reloaded.